feat: added removal of /var/run/reboot-required if needed
If /var/run/reboot-required was not present when the script started but is
present when it ends it was created just for the removed kernel(s).
Since we only remove old unused kernels a reboot should not be needed.
Therefore we remove this file.
If it was present when the script started we keep it since some other package
created it.